Moved error to separate module
